add SONOFF SA-028 to definition of Z111PL0H-1JX and move from woolley.js to sonoff.js
